Add testing against local asset server
Note that the definition of the `AssetsServer` is currently duplicated within `test_helper.exs`. This is because the desired definition at `test/support/assets_server.ex` is not loading in the tests, for some reason. That needs to be fixed.

An auxiliary addition is `ConsoleMessage`, in reaction to the matching newly sent data.