Add capturing of top-level fields
When creating resources, the optional "extra" fields are now attached to the created struct, if found on `args.initializer`.

Relatedly, those extra fields are specified as a list of keys rather than a keyword list with `nil` values. This makes the whole creation process cleaner.

Note that our test asserts that Chromium major version is >= 90, which is somewhat at odds with the version handling in `BrowserType.connect/1`.