Add `Config` module
There are a large number of configuration params and options in the Playwright ecosystem, only a small handful of which have been implemented here. As the list is starting to grow, it's time to organize those better, following the patterns in the extant implementations.

Introducing this module also provides a nice place for adding documentation for configuration.

Also, the asset server is not something that would be used outside of testing while contributing *to this project*. So, the the environment variable is still available, but there is no longer a :playwright config option.