introducing: `Channel.Command`
The `Command` module replaces ChannelMessage and moves the `:id` incrementing into `.new` on that new module, as it felt a little weird to have that responsibility be outside of the module.

This change also removes the notion of "locals": state that is held locally, pending a response with which to be merged. The intention is that this should be handled by way of some sort of response callback handler, or similar. For now, the only use was with Context <-> Page, so it's being handled as a somewhat brutish one-off.

This is all part of breaking up Connection.