Reworking internals and test structure (WIP)
We're attempting to transition away for the intermediate `channel__on` approach. To that end, the event/method handling and management are being built up.

The code at `Channel.on(connection, {"navigated", instance}, &handle_event/2)` and its neighbors make for good examples a step toward this transition.

Some set of errors, in response to commands/messages, are now handled and published (e.g., `Page.goto` handles some RuntimeErrors and raises appropriately).

The top-most `Channel.Event.handle/...` is a start toward what should be a cleaner, simpler, and more standard implementation.

There are a couple of nice examples of using `Task` to good effect.

Setting up to do some filesystem organization for the test files (and module renaming to match).

`NetworkEventTest` is the test from which to continue.