Redesign `:fields` to `@property` * n
This change redesigns the mechanism by with a `ChannelOwner` (using) module specifies properties that should be extracted from `:initializer` and made available on the resource. It also generates a sort of "getter" function for the property.

The idea is that this should open up possibilities for better documentation and additional macro-time capabilities.

See `Page.url/1` for a nice example of how the `@property` macro results in a "getter" (i.e., `Page.main_frame/1`).