Bundle/config Playwright node CLI
The Playwright node.js CLI is bundled via `esbuild` and shipped with playwright-elixir in `priv/static`.

`ws` is a runtime dependency that cannot be bundled as normal... it is referenced via `require.resolve('ws')`, which resolves to a path on the filesystem.

This change also moves the `playwright_cli_path` config from `ConnectOptions` to `LaunchOptions`.

Co-Authored-By: Greg Woodward <211737+woodward@users.noreply.github.com>