check: v1.42.0 - BIG (temp?) changes
Fundamental changes within playwright-core are proving difficult to match for
the driver and browser installation tooling. While undesireable, for the
timebeing, the entire nodejs package tree is being vendored into `priv/static`.
The aim is to continue making progress toward upgrading to the latest
playwright-core. Once that's done, we're revisit the vendoring approach.

This change also removes Electron from the package, as it is massive and not
currently covered, exercised, or supported.